kirillov made db structure

+++ ML.DataRepository/Models/AnalogSignalLog.cs
@@ -7,7 +7,7 @@
         public int BlockLogId { get; set; }
         public int SignalTypeId { get; set; }
 
-        public int SignalValue { get; set; }
+        public double SignalValue { get; set; }
 
         public virtual BlockLog BlockLog { get; set; }
         public virtual AnalogSignal SignalType { get; set; }
